split extract_demographic_factors from model_data_extract
